Refactored Fileconfig to move into api
